requirements.txt
torch
numpy
Pillow
tqdm
python-chess
